make ColumnViewer work with ConstColumn
