+++ common/functions/src/scalars/conditionals/conditional.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::scalars::function_factory::FunctionFactory;
+use crate::scalars::Function2Factory;
 use crate::scalars::IfFunction;
 
 #[derive(Clone)]
 pub struct ConditionalFunction;
 
 impl ConditionalFunction {
-    pub fn register(factory: &mut FunctionFactory) {
-        // factory.register("if", IfFunction::desc());
+    pub fn register(factory: &mut Function2Factory) {
+        factory.register("if", IfFunction::desc());
     }
 }
