--- common/functions/src/scalars/conditionals/conditional2.rs
@@ -1,25 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use crate::scalars::Function2Factory;
-use crate::scalars::IfFunction2;
-
-#[derive(Clone)]
-pub struct ConditionalFunction2;
-
-impl ConditionalFunction2 {
-    pub fn register2(factory: &mut Function2Factory) {
-        factory.register("if", IfFunction2::desc());
-    }
-}
