+++ common/functions/src/scalars/function_factory.rs
@@ -23,7 +23,6 @@ use once_cell::sync::Lazy;
 use super::Function2Convertor;
 use super::Function2Factory;
 use crate::scalars::ArithmeticFunction;
-use crate::scalars::ConditionalFunction;
 use crate::scalars::DateFunction;
 use crate::scalars::Function;
 use crate::scalars::LogicFunction;
@@ -156,7 +155,6 @@ static FUNCTION_FACTORY: Lazy<Arc<FunctionFactory>> = Lazy::new(|| {
     NullableFunction::register(&mut function_factory);
     StringFunction::register(&mut function_factory);
     UdfFunction::register(&mut function_factory);
-    ConditionalFunction::register(&mut function_factory);
     DateFunction::register(&mut function_factory);
     OtherFunction::register(&mut function_factory);
     MathsFunction::register(&mut function_factory);
